Add BlogContent;

+++ src/main/java/impl/BlogContentDaoImpl.java
@@ -24,7 +24,7 @@ public class BlogContentDaoImpl extends BaseDao<BlogContent> implements BlogCont
     }
 
     @Override
-    public List<BlogContent> ShowContent(BigDecimal i) {
+    public List<BlogContent> ShowContent(int i) {
         return null;
     }
 }
